+++ src/main/java/com/soptie/server/memberRoutine/controller/v1/MemberHappinessRoutineController.java
@@ -1,7 +1,8 @@
-package com.soptie.server.memberRoutine.controller.happiness;
+package com.soptie.server.memberRoutine.controller.v1;
 
 import com.soptie.server.common.dto.BaseResponse;
 import com.soptie.server.common.dto.SuccessResponse;
+import com.soptie.server.memberRoutine.controller.v1.api.MemberHappinessRoutineApi;
 import com.soptie.server.memberRoutine.controller.v1.dto.request.MemberHappinessRoutineRequest;
 import com.soptie.server.memberRoutine.controller.v1.dto.response.MemberHappinessRoutineResponse;
 import com.soptie.server.memberRoutine.service.MemberHappinessRoutineService;
